backend/requirements.txt
flask==2.3.2
flask-sqlalchemy==3.0.5
flask-cors==4.0.0
pymysql==1.0.3
bcrypt==4.0.1
python-dotenv==1.0.0
flask_httpauth
pyjwt
flask-migrate==4.0.5  # 升级版本